requirements.txt
aliyun-python-sdk-core==2.13.5
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
Django==2.0
django-debug-toolbar==2.0
django-haystack==2.8.1
django-shortuuidfield==0.1.3
djangorestframework==3.7.3
idna==2.6
jieba==0.39
jmespath==0.9.4
mysqlclient==1.4.2.post1
olefile==0.46
Pillow==6.1.0
PyMySQL==0.9.3
python3-memcached==1.51
pytz==2019.1
qiniu==7.2.4
redis==3.3.5
requests==2.18.4
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.22
webencodings==0.5.1
Whoosh==2.7.4
